refactor: reorganize layout components
- Move Sidebar components to src/components/layout/
- Move NotFound to src/components/common/
- Create barrel exports for layout and common components
- Update imports in _layout.tsx and __root.tsx
- Better organization with clear separation of concerns
