+++ frontend/src/components/Common/index.ts
@@ -1,0 +1,1 @@
+export { default as NotFound } from './NotFound'
