+++ frontend/src/components/layout/SidebarItems.tsx
