+++ frontend/src/components/layout/index.ts
@@ -1,0 +1,2 @@
+export { default as Sidebar } from './Sidebar'
+export { default as SidebarItems } from './SidebarItems'
