+++ frontend/src/routes/_layout.tsx
@@ -1,7 +1,7 @@
 import { Flex } from "@chakra-ui/react"
 import { Outlet, createFileRoute, redirect } from "@tanstack/react-router"
 
-import Sidebar from "@/components/Common/Sidebar"
+import { Sidebar } from "@/components/layout"
 import { isAuthenticated } from "@/lib/api/client"
 
 export const Route = createFileRoute("/_layout")({
